Extract safe_stop_emit.rs from turn.rs (parent #680)
Move the Issue #654 `agent.safe_stop.report` emit cluster out of
`turn.rs` into a focused sibling module so `turn.rs` keeps shrinking
toward responsibility-focused units (parent meta-issue #680).

Eight production helpers + one private collector extracted as free
functions taking `&mut Agent` / `&Agent` (matching `actor_loop_flow` /
`reply_retry` / earlier vertical-slice precedent):

- `record_safe_stop_report` (Task D.4 chokepoint: per-StopReason dedup
  → emit 4 job reports → SafeStopReport::build_from → bounded payload
  → `log_llm_event` with final `mask_payload_inplace` defence)
- `resolve_current_role_for_safe_stop` (CB-003 / Task D.5 §6.4 SSOT
  priority chain)
- 6 emit shells: `_for_diagnostic_target_missing` /
  `_for_verifier_failed_safe_stop` / `_for_verifier_weak`
  (#[cfg(test)]) / `_for_artifact_completion_failed` /
  `_for_verifier_missing` / `_for_repair_terminal`
- `emit_repair_safe_stop_report` (shared repair-job-driven emit
  helper)
- `collect_owned_test_artifacts` (Task D.6 Owned-validated test
  artifact collector, ≤ 8 paths, `classify_ownership` gated)

Behavior unchanged: same per-StopReason dedup via
`Agent.safe_stop_report_emitted`, same job-report-then-safe-stop emit
order (Issue #666 CB-001 fix invariant), same SafeStopContext shape,
same §6.4 priority chain (explicit role → semantic_plan → recovery
target → target_hint → reconstructed TaskContract), same masked
final-defence emit through `log_llm_event`.

`pub(super)` limited / no facade re-export (DR3-001). Callers updated
in `turn.rs` (2 self-callers), `verifier_orchestration.rs` (3 sites
including `emit_safe_stop_report_for_repair_exhausted` helper),
`repair_job_dispatch.rs` (5 sites), `actor_loop_flow.rs` (4 sites),
`turn_tests.rs` (6 sites), and `loop_run.rs` (5 #[cfg(test)] test
seams) to the free-fn form.

LOC delta:
- turn.rs: 2,752 → 2,397 (-355)
- new module: +376

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
