+++ src/agent/loop_run/turn.rs
@@ -23,7 +23,6 @@ use super::file_excerpt::{
 use super::path_helpers::normalize_memory_path;
 use super::photon_feedback_derive::request_explicitly_requests_script_execution;
 use super::plan_mode_helpers::assistant_model_for_mode;
-use super::safe_stop_payload::{build_safe_stop_payload, collect_recent_action_labels};
 use super::small_helpers::{raw_mode_safe_text, rfc3339_now_utc, user_interrupt_result};
 use super::summary::ExitReason;
 use super::tool_display::progress_path_display;
@@ -34,8 +33,7 @@ use super::tool_history::{
     focused_edit_target_already_read, focused_read_target_for_directory,
     has_successful_non_plan_repo_edit,
     has_successful_non_plan_repo_edit_after_latest_truncated_tool_call,
-    is_preferred_read_edit_target, latest_successful_read_existing_path, latest_user_turn_slice,
-    latest_verifier_repair_note_index, recent_truncated_tool_call_attempt,
+    is_preferred_read_edit_target, latest_user_turn_slice, recent_truncated_tool_call_attempt,
 };
 use super::tool_policy::{
     EffectiveToolPolicy, EffectiveToolPolicyReason, FocusedEditPolicy,
@@ -246,7 +244,7 @@ impl Agent {
         let attempt = job.setup_attempts_used as usize;
         let attempt_limit = job.retry_budget as usize;
         if exhausted {
-            self.emit_safe_stop_report_for_verifier_missing();
+            super::safe_stop_emit::emit_safe_stop_report_for_verifier_missing(self);
             return true;
         }
         write_stdout_rendered(
@@ -267,15 +265,6 @@ impl Agent {
         false
     }
 
-    pub(super) fn emit_safe_stop_report_for_repair_terminal(
-        &mut self,
-        reason: super::repair_job::RepairTerminalReason,
-    ) {
-        let Some(stop_reason) = reason.safe_stop_reason() else {
-            return;
-        };
-        self.emit_repair_safe_stop_report(stop_reason);
-    }
     pub(super) fn current_assistant_model(&self) -> String {
         assistant_model_for_mode(
             self.session.mode_state.mode,
@@ -827,351 +816,7 @@ impl Agent {
         // role / expected target / actual actions / hypothesis ledger in a
         // single structured event. Per-StopReason dedup is enforced by
         // `record_safe_stop_report`.
-        self.emit_safe_stop_report_for_diagnostic_target_missing();
-    }
-
-    /// Issue #654 / CB-003 — `current_role` SSOT fallback for safe-stop
-    /// report emission. Resolves the per-§6.4 priority order so all five
-    /// emit paths (`diagnostic_target_missing`, `verifier_failed_safe_stop`,
-    /// `verifier_weak`, `verifier_missing`, `artifact_completion_failed`)
-    /// reach into the same single source of truth.
-    ///
-    /// Priority order (first hit wins; `None` only when every source is
-    /// silent):
-    /// 1. `repair_job.semantic_plan.preferred_repair_role`
-    ///    — the verifier-driven semantic planner's authoritative pick.
-    /// 2. `current_artifact_recovery_target.role`
-    ///    — the selected artifact recovery target the planner is currently
-    ///    driving toward.
-    /// 3. `repair_job.target_hint.role`
-    ///    — the explicit repair target hint carried on `RepairJob`.
-    /// 4. The first `required_artifact` of the active `TaskContract`
-    ///    reconstructed from the current request text.
-    ///
-    /// `explicit_role` allows the `artifact_completion_failed` path to
-    /// preempt the priority order with the role the host emit point already
-    /// has in hand (the role the retry budget exhausted on).
-    pub(super) fn resolve_current_role_for_safe_stop(
-        &self,
-        explicit_role: Option<super::task_contract::ArtifactRole>,
-    ) -> Option<super::task_contract::ArtifactRole> {
-        if let Some(role) = explicit_role {
-            return Some(role);
-        }
-        if let Some(job) = self.repair_job.as_ref()
-            && let Some(plan) = job.semantic_plan.as_ref()
-        {
-            return Some(plan.preferred_repair_role);
-        }
-        if let Some(target) = self.current_artifact_recovery_target.as_ref() {
-            return Some(target.role);
-        }
-        if let Some(job) = self.repair_job.as_ref()
-            && let Some(hint) = job.target_hint.as_ref()
-        {
-            return Some(hint.role);
-        }
-        let request = self.active_request_text().unwrap_or_default();
-        if !request.is_empty() {
-            let contract = super::task_contract::TaskContract::from_request(&request);
-            if let Some(role) = contract.required_artifacts.first().copied() {
-                return Some(role);
-            }
-        }
-        None
-    }
-
-    /// Issue #654 — diagnostic_target_missing path: build a
-    /// `SafeStopInput::FromRepair { DiagnosticTargetMissing }` and forward to
-    /// `record_safe_stop_report`. Pulled into a separate method so the
-    /// orchestration of context-collection from `Agent` state stays out of
-    /// the lifecycle of `record_verifier_diagnostic_unavailable`.
-    pub(super) fn emit_safe_stop_report_for_diagnostic_target_missing(&mut self) {
-        let Some(job) = self.repair_job.clone() else {
-            return;
-        };
-        let session_id = self.session_store.session_id().to_string();
-        let turn_index = self.current_turn_index as u64;
-        let scope = super::task_workspace_scope::TaskWorkspaceScope::detect(
-            &self.work_root,
-            self.active_request_text().unwrap_or_default().as_str(),
-        );
-        let candidates: Vec<String> = job
-            .changed_file_hints
-            .iter()
-            .map(|hint| hint.path.clone())
-            .collect();
-        let latest_read = latest_successful_read_existing_path(
-            &self.session.messages,
-            &self.work_root,
-            latest_verifier_repair_note_index(&self.session.messages),
-        );
-        let expected = job
-            .target_hint
-            .as_ref()
-            .map(|hint| std::path::PathBuf::from(&hint.path));
-        // CB-003: route through the §6.4 SSOT fallback chain so the
-        // diagnostic_target_missing path is not limited to semantic_plan.
-        let current_role = self.resolve_current_role_for_safe_stop(None);
-        let input = super::repair_job::SafeStopInput::FromRepair {
-            job: &job,
-            stop_reason: super::repair_job::StopReason::DiagnosticTargetMissing,
-            owned_test_artifacts: Vec::new(),
-        };
-        let ctx = super::repair_job::SafeStopContext {
-            current_role,
-            expected_target: expected.as_deref(),
-            actual_actions_raw: collect_recent_action_labels(&self.session.messages),
-            latest_successful_read: latest_read.as_deref(),
-            task_workspace_scope: &scope,
-            candidates,
-            session_id: &session_id,
-            turn_index,
-        };
-        self.record_safe_stop_report(input, ctx);
-    }
-
-    /// Issue #654 (Task D.4) — thin shell that dedups per StopReason, builds
-    /// the structured report via `SafeStopReport::build_from`, renders the
-    /// bounded payload via `build_safe_stop_payload`, and emits the
-    /// `agent.safe_stop.report` event through `log_llm_event` (which routes
-    /// through `mask_payload_inplace` — the final defense line).
-    pub(super) fn record_safe_stop_report(
-        &mut self,
-        input: super::repair_job::SafeStopInput<'_>,
-        ctx: super::repair_job::SafeStopContext<'_>,
-    ) {
-        let stop_reason = input.stop_reason();
-        if self.safe_stop_report_emitted.contains(&stop_reason) {
-            return;
-        }
-        // Issue #666 (CB-001 fix): emit per-turn job reports BEFORE the
-        // SafeStopReport so the llm-io event order is
-        // `agent.{x}.report → agent.safe_stop.report` per design Section
-        // 8-2. Per-turn dedup in `maybe_emit_job_reports` makes the
-        // post-`run_turn` finalizer at `handle_user_message` a no-op for
-        // any kind already emitted here. We record the stop_reason in
-        // `safe_stop_report_emitted` AFTER `maybe_emit_job_reports` so
-        // the SafeStopLinkage built into the job reports observes the
-        // pre-stop state of the dedup set (`report_emitted=false`); the
-        // job reports still carry the actual stop_reason via
-        // `safe_stop.reason` once the post-run finalizer is dedup'd out.
-        // Order: build linkage → emit 4 job reports → emit safe_stop.
-        let linkage_reason = Some(stop_reason.as_str().to_string());
-        self.maybe_emit_job_reports_with_linkage(linkage_reason);
-        let report = super::repair_job::SafeStopReport::build_from(input, ctx);
-        let payload = build_safe_stop_payload(&report);
-        log_llm_event("agent.safe_stop.report", payload);
-        self.safe_stop_report_emitted.insert(stop_reason);
-    }
-
-    /// Issue #654 (E.3) — `verifier_failed_safe_stop` emit shell. Builds a
-    /// `FromRepair { VerifierFailedSafeStop }` input from `self.repair_job`.
-    pub(super) fn emit_safe_stop_report_for_verifier_failed_safe_stop(&mut self) {
-        self.emit_repair_safe_stop_report(super::repair_job::StopReason::VerifierFailedSafeStop);
-    }
-
-    #[cfg(test)]
-    pub(super) fn emit_safe_stop_report_for_verifier_weak(&mut self) {
-        self.emit_repair_safe_stop_report(super::repair_job::StopReason::VerifierWeak);
-    }
-
-    /// Issue #654 (E.2) — `artifact_completion_failed` emit shell. Unlike the
-    /// other shells, this path can fire BEFORE a verifier-driven `RepairJob`
-    /// has been built (the role-specific retry budget exhausts during pure
-    /// artifact-completion attempts). When `self.repair_job` is `None` we
-    /// fall back to a synthetic empty `RepairJob` whose only meaningful field
-    /// is `target_hint` (carried from the explicit role + path the caller
-    /// supplies) so the emitted payload still carries `current_role` and
-    /// `expected_target` for downstream `/bug-fix` consumers.
-    pub(super) fn emit_safe_stop_report_for_artifact_completion_failed(
-        &mut self,
-        role: super::task_contract::ArtifactRole,
-        expected_target_path: Option<String>,
-    ) {
-        if self
-            .safe_stop_report_emitted
-            .contains(&super::repair_job::StopReason::ArtifactCompletionFailed)
-        {
-            return;
-        }
-        let session_id = self.session_store.session_id().to_string();
-        let turn_index = self.current_turn_index as u64;
-        let scope = super::task_workspace_scope::TaskWorkspaceScope::detect(
-            &self.work_root,
-            self.active_request_text().unwrap_or_default().as_str(),
-        );
-        let owned_test_artifacts = self.collect_owned_test_artifacts();
-        // Prefer the real RepairJob (when one exists, e.g. the RepairArtifact
-        // exhaustion path at the verifier-repair stage) so failure_signature /
-        // command / output_excerpt remain accurate. Otherwise fall back to a
-        // synthetic shell whose only carried context is the target path the
-        // caller provided.
-        let job = match self.repair_job.clone() {
-            Some(job) => job,
-            None => {
-                let mut shell = super::repair_job::RepairJob::empty_synthetic();
-                if let Some(path) = expected_target_path.clone() {
-                    shell.target_hint = Some(super::task_contract::RecoveryTargetHint {
-                        role,
-                        path,
-                        reason: "artifact_completion_failed".to_string(),
-                    });
-                }
-                shell
-            }
-        };
-        let expected = expected_target_path
-            .map(std::path::PathBuf::from)
-            .or_else(|| {
-                job.target_hint
-                    .as_ref()
-                    .map(|hint| std::path::PathBuf::from(&hint.path))
-            });
-        // CB-003: still preempt with the explicit role the caller supplied
-        // (the artifact-completion retry budget exhausted on this role), but
-        // route the no-explicit-role branches through the §6.4 helper so the
-        // priority order (semantic_plan -> recovery target -> target_hint ->
-        // contract) stays single-sourced.
-        let current_role = self.resolve_current_role_for_safe_stop(Some(role));
-        let candidates: Vec<String> = job
-            .changed_file_hints
-            .iter()
-            .map(|hint| hint.path.clone())
-            .collect();
-        let input = super::repair_job::SafeStopInput::FromRepair {
-            job: &job,
-            stop_reason: super::repair_job::StopReason::ArtifactCompletionFailed,
-            owned_test_artifacts,
-        };
-        let ctx = super::repair_job::SafeStopContext {
-            current_role,
-            expected_target: expected.as_deref(),
-            actual_actions_raw: collect_recent_action_labels(&self.session.messages),
-            latest_successful_read: None,
-            task_workspace_scope: &scope,
-            candidates,
-            session_id: &session_id,
-            turn_index,
-        };
-        self.record_safe_stop_report(input, ctx);
-    }
-
-    /// Shared helper for repair-job-driven emit paths (E.2 / E.3 / E.4).
-    pub(super) fn emit_repair_safe_stop_report(
-        &mut self,
-        stop_reason: super::repair_job::StopReason,
-    ) {
-        let Some(job) = self.repair_job.clone() else {
-            return;
-        };
-        let session_id = self.session_store.session_id().to_string();
-        let turn_index = self.current_turn_index as u64;
-        let scope = super::task_workspace_scope::TaskWorkspaceScope::detect(
-            &self.work_root,
-            self.active_request_text().unwrap_or_default().as_str(),
-        );
-        let owned_test_artifacts = self.collect_owned_test_artifacts();
-        let expected = job
-            .target_hint
-            .as_ref()
-            .map(|hint| std::path::PathBuf::from(&hint.path));
-        // CB-003: route through the §6.4 SSOT fallback chain.
-        let current_role = self.resolve_current_role_for_safe_stop(None);
-        let candidates: Vec<String> = job
-            .changed_file_hints
-            .iter()
-            .map(|hint| hint.path.clone())
-            .collect();
-        let input = super::repair_job::SafeStopInput::FromRepair {
-            job: &job,
-            stop_reason,
-            owned_test_artifacts,
-        };
-        let ctx = super::repair_job::SafeStopContext {
-            current_role,
-            expected_target: expected.as_deref(),
-            actual_actions_raw: collect_recent_action_labels(&self.session.messages),
-            latest_successful_read: None,
-            task_workspace_scope: &scope,
-            candidates,
-            session_id: &session_id,
-            turn_index,
-        };
-        self.record_safe_stop_report(input, ctx);
-    }
-
-    /// Issue #654 (E.5) — `verifier_missing` emit shell. Uses the
-    /// `FromMissingVerifier` builder so empty `failure_signature` fallback
-    /// detection by downstream consumers does not misfire (R8).
-    pub(super) fn emit_safe_stop_report_for_verifier_missing(&mut self) {
-        let Some(job) = self.missing_verifier_job.clone() else {
-            return;
-        };
-        let session_id = self.session_store.session_id().to_string();
-        let turn_index = self.current_turn_index as u64;
-        let scope = super::task_workspace_scope::TaskWorkspaceScope::detect(
-            &self.work_root,
-            self.active_request_text().unwrap_or_default().as_str(),
-        );
-        let owned_test_artifacts = self.collect_owned_test_artifacts();
-        let input = super::repair_job::SafeStopInput::FromMissingVerifier {
-            job: &job,
-            owned_test_artifacts,
-        };
-        // CB-003: verifier_missing has no `RepairJob`, so the helper falls
-        // through to current_artifact_recovery_target -> reconstructed
-        // TaskContract.required_artifacts.first() for downstream `/bug-fix`.
-        let current_role = self.resolve_current_role_for_safe_stop(None);
-        let ctx = super::repair_job::SafeStopContext {
-            current_role,
-            expected_target: None,
-            actual_actions_raw: collect_recent_action_labels(&self.session.messages),
-            latest_successful_read: None,
-            task_workspace_scope: &scope,
-            candidates: Vec::new(),
-            session_id: &session_id,
-            turn_index,
-        };
-        self.record_safe_stop_report(input, ctx);
-    }
-
-    /// Issue #654 (DR3-002 / Task D.6) — collect Owned-validated test artifact
-    /// relative paths from the current turn's edits, gated by
-    /// `classify_ownership`. Returns at most `SAFE_STOP_OWNED_TEST_ARTIFACTS_MAX`
-    /// paths; the builder re-applies syntactic safety as defense-in-depth.
-    fn collect_owned_test_artifacts(&self) -> Vec<String> {
-        let scope = super::task_workspace_scope::TaskWorkspaceScope::detect(
-            &self.work_root,
-            self.active_request_text().unwrap_or_default().as_str(),
-        );
-        let mut out: Vec<String> = Vec::new();
-        for rel in self.turn_edited_relative_paths.iter() {
-            // Only test files qualify.
-            if !crate::util::file_classify::is_test_file(std::path::Path::new(rel)) {
-                continue;
-            }
-            let inputs = super::artifact_ownership::OwnershipInputs {
-                work_root: &self.work_root,
-                relative_path: rel.as_str(),
-                scope: &scope,
-                edited_this_session: true,
-                scaffold_changed: false,
-                verifier_passed_in_scope: false,
-                // Issue #661 (Task 3.1): legacy helper — not one of the 4
-                // verifier-path SSOT sites. Iteration-3 may revisit if this
-                // shadow consumer needs verifier-binding semantics.
-                nested_test_admission: super::artifact_ownership::NestedTestAdmission::default(),
-            };
-            if super::artifact_ownership::classify_ownership(inputs)
-                == super::artifact_ownership::ArtifactOwnership::Owned
-            {
-                out.push(rel.clone());
-            }
-        }
-        out.sort();
-        out.truncate(8);
-        out
+        super::safe_stop_emit::emit_safe_stop_report_for_diagnostic_target_missing(self);
     }
 
     pub(super) fn verifier_repair_pass_wall_clock_timeout_error(
